Make player entry of game patterns functionally complete.
Game will now reset if a mistake is made in strict mode. If not in
strict mode, game will repeat the pattern until player gets it right.

Game buttons also now light up as long as player has the mouse pressed
over the top of the button. Verification of proper button press has
been modified to account for this.

Highlighting and dragging of html elements has been disabled to improve
both user experience and functionally in case the user tries to drag
a game play button or highlight random text.